requirements.txt
apache-airflow==2.8.1
pandas
pyarrow
google-cloud-storage
google-cloud-bigquery
dbt-bigquery
requests